requirements.txt
pyautogui
pillow
pytesseract
rapidfuzz
python-Levenshtein
# Streamlit web app dependencies
streamlit
pandas
plotly
# GUI dependencies for coordinate_adjuster.py
# Note: tkinter comes with most Python installations
# If tkinter is not available, use: brew install python-tk
